Friendbot, comments: moved services/horizon/render back to services/horizon/internal/render
