+++ services/horizon/internal/actions/helpers.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/ledger"
 	"github.com/stellar/go/services/horizon/internal/toid"
-	"github.com/stellar/go/services/horizon/render/problem"
+	"github.com/stellar/go/services/horizon/internal/render/problem"
 	"github.com/stellar/go/strkey"
 	"github.com/stellar/go/support/errors"
 	"github.com/stellar/go/support/time"
