+++ services/horizon/internal/actions/helpers_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stellar/go/services/horizon/internal/ledger"
 	"github.com/stellar/go/services/horizon/internal/test"
 	"github.com/stellar/go/services/horizon/internal/toid"
-	"github.com/stellar/go/services/horizon/render/problem"
+	"github.com/stellar/go/services/horizon/internal/render/problem"
 	"github.com/stellar/go/xdr"
 	"github.com/zenazn/goji/web"
 )
