+++ services/horizon/internal/actions/responders.go
@@ -1,6 +1,6 @@
 package actions
 
-import "github.com/stellar/go/services/horizon/render/sse"
+import "github.com/stellar/go/services/horizon/internal/render/sse"
 
 // JSON implementors can respond to a request whose response type was negotiated
 // to be MimeHal or MimeJSON.
