+++ services/horizon/internal/actions_metrics.go
@@ -2,7 +2,7 @@ package horizon
 
 import (
 	"github.com/rcrowley/go-metrics"
-	"github.com/stellar/go/services/horizon/render/hal"
+	"github.com/stellar/go/services/horizon/internal/render/hal"
 )
 
 // MetricsAction collects and renders a snapshot from the metrics system that
