+++ services/horizon/internal/actions_not_found.go
@@ -1,6 +1,6 @@
 package horizon
 
-import "github.com/stellar/go/services/horizon/render/problem"
+import "github.com/stellar/go/services/horizon/internal/render/problem"
 
 // NotFoundAction renders a 404 response
 type NotFoundAction struct {
