+++ services/horizon/internal/actions_not_implemented.go
@@ -1,6 +1,6 @@
 package horizon
 
-import "github.com/stellar/go/services/horizon/render/problem"
+import "github.com/stellar/go/services/horizon/internal/render/problem"
 
 // NotImplementedAction renders a NotImplemented prblem
 type NotImplementedAction struct {
