+++ services/horizon/internal/actions_path.go
@@ -3,7 +3,7 @@ package horizon
 import (
 	"github.com/stellar/go/services/horizon/internal/paths"
 	"github.com/stellar/go/services/horizon/internal/resource"
-	"github.com/stellar/go/services/horizon/render/hal"
+	"github.com/stellar/go/services/horizon/internal/render/hal"
 )
 
 // PathIndexAction provides path finding
