+++ services/horizon/internal/actions_rate_limit_exceeded.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/zenazn/goji/web"
 
-	"github.com/stellar/go/services/horizon/render/problem"
+	"github.com/stellar/go/services/horizon/internal/render/problem"
 )
 
 // RateLimitExceededAction renders a 429 response
