+++ services/horizon/internal/actions_trade.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/resource"
-	"github.com/stellar/go/services/horizon/render/hal"
+	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"github.com/stellar/go/support/time"
 	"github.com/stellar/go/xdr"
 )
