+++ services/horizon/internal/app.go
@@ -22,7 +22,7 @@ import (
 	"github.com/stellar/go/services/horizon/internal/paths"
 	"github.com/stellar/go/services/horizon/internal/reap"
 	"github.com/stellar/go/services/horizon/internal/txsub"
-	"github.com/stellar/go/services/horizon/render/sse"
+	"github.com/stellar/go/services/horizon/internal/render/sse"
 	"github.com/stellar/go/support/db"
 	"golang.org/x/net/context"
 	"golang.org/x/net/http2"
