+++ services/horizon/internal/render/sse/main.go
