+++ services/horizon/internal/render/sse/stream.go
