+++ services/horizon/internal/resource/account.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stellar/go/services/horizon/httpx"
 	"github.com/stellar/go/services/horizon/internal/db2/core"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
-	"github.com/stellar/go/services/horizon/render/hal"
+	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"golang.org/x/net/context"
 )
