+++ services/horizon/internal/resource/effects/main.go
@@ -3,7 +3,7 @@ package effects
 import (
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/resource/base"
-	"github.com/stellar/go/services/horizon/render/hal"
+	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"golang.org/x/net/context"
 )
