+++ services/horizon/internal/resource/trade.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stellar/go/amount"
 	"github.com/stellar/go/services/horizon/httpx"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
-	"github.com/stellar/go/services/horizon/render/hal"
+	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"golang.org/x/net/context"
 )
