+++ services/horizon/internal/resource/transaction.go
@@ -8,7 +8,7 @@ import (
 	"github.com/guregu/null"
 	"github.com/stellar/go/services/horizon/httpx"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
-	"github.com/stellar/go/services/horizon/render/hal"
+	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"golang.org/x/net/context"
 )
